Rename and update sample
