+++ samples/show-device-location-with-nmea-data-sources/src/main/AndroidManifest.xml
@@ -6,12 +6,12 @@
     <application><activity
             android:name=".DownloadActivity"
             android:exported="true"
-            android:label="@string/display_device_location_with_nmea_data_sources_app_name">
+            android:label="@string/show_device_location_with_nmea_data_sources_app_name">
         </activity>
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/display_device_location_with_nmea_data_sources_app_name">
+            android:label="@string/show_device_location_with_nmea_data_sources_app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
             </intent-filter>
