+++ samples/show-device-location-with-nmea-data-sources/src/main/res/drawable/ic_round_pause_24.xml
