+++ samples/show-device-location-with-nmea-data-sources/src/main/res/drawable/ic_round_play_arrow_24.xml
