+++ samples/show-device-location-with-nmea-data-sources/src/main/res/layout/show_device_location_with_nmea_data_sources_activity_main.xml
