+++ samples/show-device-location-with-nmea-data-sources/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="display_device_location_with_nmea_data_sources_app_name">Display device location with NMEA data sources</string>
+    <string name="show_device_location_with_nmea_data_sources_app_name">Show device location with NMEA data sources</string>
     <string name="start_stop_button">Start/Stop button</string>
     <string name="accuracy">Accuracy: </string>
     <string name="satellite_count">Satellite Count: </string>
